+++ sandbox/TestChroma.py
@@ -1,6 +1,6 @@
-from agentforge.utils.ChromaUtils import ChromaUtils
+from agentforge.storage.chroma_storage import ChromaStorage
 
-storage = ChromaUtils()
+storage = ChromaStorage()
 print('hi')
 
 var = storage.search_storage_by_threshold("Actions","search web", threshold=0.8, num_results=5)
