+++ sandbox/TestLLM.py
@@ -1,6 +1,4 @@
 from CustomAgents.DocsAgent import DocsAgent
-from agentforge.utils.ChromaUtils import ChromaUtils
-from agentforge.utils.function_utils import Logger
 
 docs_agent = DocsAgent()
 # kb = ChromaUtils()
